Afrigric/requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.0
xgboost==1.7.6
joblib==1.3.2
tensorflow==2.13.0
Pillow==10.0.0
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.10